chore: combine axestuple
